Split out xivmath, try to fix build
