+++ .github/workflows/backend_docker.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Build
         run: |
           npm install
-          npm install --workspaces
+          npm install --workspaces || npm install --workspaces
 
       - name: Inject commit info
         run: |
