+++ .github/workflows/frontend_gh_pages.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Build
         run: |
           npm install
-          npm install --workspaces
+          npm install --workspaces || npm install --workspaces
           npm run --workspace=gearplan-frontend build
           npm run --workspace=gearplan-frontend less
           npm run --workspace=gearplan-frontend test
