+++ .github/workflows/frontend_test_only.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Build
         run: |
           npm install
-          npm install --workspaces
+          npm install --workspaces || npm install --workspaces
           npm run --workspace=gearplan-frontend build
           npm run --workspace=gearplan-frontend less
           npm run --workspace=gearplan-frontend test
