+++ .idea/runConfigurations/build_dev.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="build dev" type="js.build_tools.npm">
-    <package-json value="$PROJECT_DIR$/package.json" />
+    <package-json value="$PROJECT_DIR$/packages/frontend/package.json" />
     <command value="run" />
     <scripts>
       <script value="build" />
