+++ packages/frontend/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "dependencies": {
     "@webcomponents/webcomponentsjs": "^2.8.0",
-    "xivmath": "../xivmath/src"
+    "xivmath": "../xivmath/src/"
   },
   "devDependencies": {
     "@types/expect": "^24.3.0",
