+++ packages/frontend/src/scripts/components/totals_display.ts
@@ -1,6 +1,6 @@
 import {CharacterGearSet} from "../gear";
 import {ChanceStat, GcdStat, MultiplierStat, RawStatKey, TickStat} from "../geartypes";
-import {NORMAL_GCD, STAT_ABBREVIATIONS, STAT_DISPLAY_ORDER} from "../xivconstants";
+import {NORMAL_GCD, STAT_ABBREVIATIONS, STAT_DISPLAY_ORDER} from "xivmath/xivconstants";
 
 export class SetViewToolbar extends HTMLElement {
