+++ packages/frontend/src/scripts/main.ts
@@ -3,7 +3,7 @@ import {SetExport, SheetExport} from "./geartypes";
 import {quickElement} from "./components/util";
 import {getShortLink} from "./external/shortlink_server";
 import {getBisSheet} from "./external/static_bis";
-import {JobName} from "./xivconstants";
+import {JobName} from "xivmath/xivconstants";
 import {LoadingBlocker} from "./components/loader";
 import {earlyEmbedInit, openEmbed} from "./embed";
 import {SETTINGS} from "./persistent_settings";
