+++ packages/frontend/src/scripts/sims/rpr_sheet_sim.ts
@@ -1,5 +1,5 @@
 import {SimSettings, SimSpec} from "../simulation";
-import {STANDARD_ANIMATION_LOCK} from "../xivconstants";
+import {STANDARD_ANIMATION_LOCK} from "xivmath/xivconstants";
 import {ArcaneCircle, DeathsDesign} from "./buffs";
 import {BaseMultiCycleSim, CycleProcessor, CycleSimResult, ExternalCycleSettings, Rotation} from "./sim_processors";
 import {Ability, GcdAbility, OgcdAbility} from "./sim_types";
