+++ packages/frontend/src/scripts/sims/sge_sheet_sim.ts
@@ -1,6 +1,6 @@
 import {SimResult, SimSettings, SimSpec, Simulation} from "../simulation";
 import {CharacterGearSet} from "../gear";
-import {applyDhCrit, baseDamage} from "../xivmath";
+import {applyDhCrit, baseDamage} from "xivmath/xivmath";
 import {ComputedSetStats} from "../geartypes";
 
 import {FieldBoundCheckBox, labeledCheckbox} from "../components/util";
