+++ packages/frontend/src/scripts/test/sims/cycle_processor_tests.ts
@@ -13,8 +13,8 @@ import {assertClose, makeFakeSet} from "../test_utils";
 import {assertSimAbilityResults, setPartyBuffEnabled, UseResult} from "./sim_test_utils";
 import {SimSettings, SimSpec} from "../../simulation";
 import {JobMultipliers} from "../../geartypes";
-import {finalizeStats} from "../../xivstats";
-import {getClassJobStats, getLevelStats} from "../../xivconstants";
+import {finalizeStats} from "xivmath/xivstats";
+import {getClassJobStats, getLevelStats} from "xivmath/xivconstants";
 import {CharacterGearSet} from "../../gear";
 import {Divination, Litany, Mug} from "../../sims/buffs";
 import {exampleGearSet} from "./common_values";
