+++ packages/xivmath/src/xivmath.ts
@@ -1,4 +1,4 @@
-import {AttackType, ComputedSetStats, JobData, JobMultipliers, LevelStats} from "./geartypes";
+import {AttackType, ComputedSetStats, JobData, JobMultipliers, LevelStats} from "../../frontend/src/scripts/geartypes";
 
 /*
     Common math for FFXIV.
